refactor console module and document items
