+++ wasmstation/src/renderer.rs
@@ -1,3 +1,5 @@
+//! Utilities for Rendering Games
+
 use std::{
     fs::{self, File},
     io::Write,
@@ -35,6 +37,7 @@ const TARGET_MS_PER_FRAME: Duration = Duration::from_millis((1000.0 / TARGET_FPS
 const SCREEN_LENGTH: usize = (SCREEN_SIZE * SCREEN_SIZE) as usize;
 const TEXTURE_LENGTH: usize = SCREEN_LENGTH * 3;
 
+/// Launch a game in a SDL2 window.
 pub fn launch(mut backend: impl Backend, path: &Path, display_scale: u32) -> anyhow::Result<()> {
     let mut save_file = path.to_path_buf();
     save_file.set_extension("disk");
