+++ wasmstation/src/utils.rs
@@ -1,7 +1,10 @@
+//! Various Utilities for WASM-4
+
 use byteorder::{LittleEndian, WriteBytesExt};
 
 use crate::wasm4::FRAMEBUFFER_SIZE;
 
+/// Returns the default WASM-4 palette.
 pub fn default_palette() -> [u8; 16] {
     let mut buf = vec![];
     buf.write_u32::<LittleEndian>(0xe0f8cf).unwrap();
@@ -12,6 +15,7 @@ pub fn default_palette() -> [u8; 16] {
     buf.try_into().expect("wrong palette size")
 }
 
+/// Returns the default WASM-4 draw colors.
 pub fn default_draw_colors() -> [u8; 2] {
     bytemuck::cast_slice(&[0x1203_u16])
         .to_vec()
@@ -19,11 +23,7 @@ pub fn default_draw_colors() -> [u8; 2] {
         .expect("wrong draw colors size")
 }
 
+/// Returns an empty WASM-4 framebuffer.
 pub fn default_framebuffer() -> [u8; FRAMEBUFFER_SIZE] {
-    (0..6400)
-        .into_iter()
-        .map(|_| 0x0000)
-        .collect::<Vec<u8>>()
-        .try_into()
-        .expect("wrong framebuffer size")
+    [0u8; 6400]
 }
